Move `numbits` parameters to end of command call
Documentation for OpenSSL suggests that this unnamed parameter should be at the end of the command call, where we had it in the middle. While this has been working fine for me, it was raised in issue #6 that this was breaking for some installations of OpenSSL.

This potentially backs the work in issue #3 to replace OpenSSL usage with a standardised Swift library - removing versions from the mix.

Have locally built authority and leaf, working fine with change. Unsure if addresses issue though as can't reproduce.